overlay: Kang parsing fix from upstream (Sprint)
Mms recieve on epic is not working due to changes
in  content disposition parsing in PduParser.java.
This change fixes the if statement which previously ran
if the preference config_mms_content_disposition_support
was true, which it always was.
This change sets the preference in the epic
overlay config.xml to be false.